backend/requirements.txt
fastapi==0.115.0
sqlmodel==0.0.22
uvicorn==0.32.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
openai==1.55.0
pydantic==2.10.0
pydantic-settings==2.7.0
alembic==1.14.0
psycopg2-binary==2.10.0
asyncpg==0.29.0
python-dotenv==1.0.1